+++ CustomCommands/Interfaces/IRegisterCommands.cs
@@ -1,6 +1,6 @@
 using CustomCommands.Model;
 
-namespace CustomCommands.Services;
+namespace CustomCommands.Interfaces;
 
 public interface IRegisterCommands
 {
